requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via typer
docker==7.0.0
    # via opendatastudio-cli (pyproject.toml)
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via opendatastudio-cli (pyproject.toml)
idna==3.7
    # via requests
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
packaging==24.0
    # via docker
pygments==2.17.2
    # via rich
requests==2.31.0
    # via docker
rich==13.7.1
    # via typer
shellingham==1.5.4
    # via typer
smmap==5.0.1
    # via gitdb
typer==0.12.3
    # via opendatastudio-cli (pyproject.toml)
typing-extensions==4.11.0
    # via typer
urllib3==2.2.1
    # via
    #   docker
    #   requests
